Amazon linux support

+++ k3s_cluster/files/k3s-install-agent.sh
@@ -1,12 +1,57 @@
 #!/bin/bash
 
-apt-get update
-apt-get install -y software-properties-common unzip
-DEBIAN_FRONTEND=noninteractive apt-get upgrade -y
-curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
-unzip awscliv2.zip
-sudo ./aws/install
-rm -rf aws awscliv2.zip
+check_os() {
+  name=$(cat /etc/os-release | grep ^NAME= | sed 's/"//g')
+  clean_name=$${name#*=}
+
+  version=$(cat /etc/os-release | grep ^VERSION_ID= | sed 's/"//g')
+  clean_version=$${version#*=}
+  major=$${clean_version%.*}
+  minor=$${clean_version#*.}
+  
+  if [[ "$clean_name" == "Ubuntu" ]]; then
+    operating_system="ubuntu"
+  elif [[ "$clean_name" == "Amazon Linux" ]]; then
+    operating_system="amazonlinux"
+  else
+    operating_system="undef"
+  fi
+
+  echo "K3S install process running on: "
+  echo "OS: $operating_system"
+  echo "OS Major Release: $major"
+  echo "OS Minor Release: $minor"
+}
+
+wait_lb() {
+while [ true ]
+do
+  curl --output /dev/null --silent -k https://${k3s_url}:6443
+  if [[ "$?" -eq 0 ]]; then
+    break
+  fi
+  sleep 5
+  echo "wait for LB"
+done
+}
+
+check_os
+
+if [[ "$operating_system" == "ubuntu" ]]; then
+  apt-get update
+  apt-get install -y software-properties-common unzip
+  DEBIAN_FRONTEND=noninteractive apt-get upgrade -y
+  curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+  unzip awscliv2.zip
+  sudo ./aws/install
+  rm -rf aws awscliv2.zip
+fi
+
+if [[ "$operating_system" == "amazonlinux" ]]; then
+  yum install -y unzip curl jq
+  aws_region="$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | jq -r .region)"
+  export AWS_DEFAULT_REGION=$aws_region
+fi
 
 k3s_install_params=()
 
@@ -31,7 +76,9 @@ K3S_VERSION=$(curl --silent https://api.github.com/repos/k3s-io/k3s/releases/lat
 K3S_VERSION="${k3s_version}"
 %{ endif }
 
+wait_lb
+
 until (curl -sfL https://get.k3s.io | INSTALL_K3S_VERSION=$K3S_VERSION K3S_TOKEN=${k3s_token} K3S_URL=https://${k3s_url}:6443 sh -s - $INSTALL_PARAMS); do
-    echo 'k3s did not install correctly'
-    sleep 2
+  echo 'k3s did not install correctly'
+  sleep 2
 done
